+++ src/main/java/com/cardonaroger/customer/CustomerDao.java
@@ -5,10 +5,10 @@ import java.util.Optional;
 
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
-    Optional<Customer> selectCustomerById(Integer id);
+    Optional<Customer> selectCustomerById(Long id);
     void insertCustomer(Customer customer);
     boolean existsPersonWithEmail(String email);
-    boolean existsPersonWithId(Integer id);
-    void deleteCustomer(Integer id);
+    boolean existsPersonWithId(Long id);
+    void deleteCustomer(Long id);
     void updateCustomer(Customer update);
 }
